+++ packages/client-reconciliation/RELEASES.md
@@ -129,3 +129,13 @@ Based on:
 - [typescript v1.0.0-alpha.10] packages/client-reconciliation
 ### Releases
 - [NPM v1.0.0-alpha.10] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.10 - packages/client-reconciliation
+
+## 2024-08-15 00:15:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.368.0 (2.399.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.11] packages/client-reconciliation
+### Releases
+- [NPM v1.0.0-alpha.11] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.11 - packages/client-reconciliation